requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
frozenlist==1.3.0
icalendar==4.0.9
idna==3.3
multidict==6.0.2
numpy==1.22.3
psycopg2==2.9.3
pycparser==2.21
pydub==0.25.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
schedule==1.1.0
six==1.16.0
SoundFile==0.10.3.post1
sqlparse==0.4.2
urllib3==1.26.9
yarl==1.7.2
